Change `failure.trace` in ci.json to be string (not array)
I understand that the original intent probably was to capture
only the stack trace with one stack frame per line
(more precisely, per array item). However, this does not work
in practice - Ruby is essentially the only target that respects this
(because it's the only language where the indidivual stack frames
are already available as an array). In other targets, the whole free-form
text stack trace is at best merely dumped as the first element of the
`trace` array, at worse it is also surrounded by other string elements
that only contain whitespace (see https://github.com/kaitai-io/ci_artifacts/blob/66e9bba/test_out/java/ci.json#L200-L204 for example -
this happened because of the `<![CDATA[` wrapping).

Therefore, it's better to reconcile with the fact that the `trace` is
a free-form text in a format specific to the target language and
testing framework and it might not only contain the raw stack
trace but can also contain error messages, lines of code, etc.
So let's change the type to a plain `string`.