Add tests {ValidEq,Expr}StrEncodings
Following the comment https://github.com/kaitai-io/kaitai_struct_ruby_runtime/pull/7#issuecomment-793155345

Test that the string parsed from stream in the specified encoding
is equal to the appropriate literal UTF-8 counterpart.

This is designed to point out a problem in Ruby, where each string
is represented as a byte array holding the characters in specified
encoding. Testing the same strings for equality will then fail if
their byte representations are not the same (i.e. if they use
encodings that represent chars differently).