PHP: use PHPUnit installed by Composer in the runtime lib dir
Using the system `phpunit` does not guarantee that you will get a
version of PHPUnit that is compatible with the desired PHP version. The
pre-installed `phpunit` on Travis CI depends on the version of the
operating system rather than the version of PHP we want to run the tests
on. So we want to ensure that we're using a known version of PHPUnit as
specified in the
[composer.json](https://github.com/kaitai-io/kaitai_struct_php_runtime/blob/master/composer.json)
of the runtime library.

This commit also resolves the `vendor/autoload.php` file from runtime
library folder in a more reliable way. This was needed for my
https://github.com/generalmimon/ks-bits-fuzzer tool to work, but it's an
improvement worth adding to the main branch as well (certainly better
than relying on a hardcoded relative path).