Regen SwitchManualIntSizeEos from KST to all but Go
Go does not support type casting yet, so the spec cannot be generated
automatically and would have to be manually patched. Moreover, the
`_raw_Body` is currently unexported and trying to access it results in a
compile error from which the partial builder cannot recover (it can only
recover from an error in the generated parser, not test spec).