junit_xml_parser: downgrade `classname` match error to warning
This is needed to fix the `aggregate/convert_to_json` crash due to error
`Unable to parse classname: "unittest.loader.ModuleImportFailure"` that
currently occurs when running ./ci-python on Python 2.

When `unittest` automatically imports modules during test discovery and
some file cannot be imported (e.g. due to a syntax error), the error is
reported under the classname `unittest.loader.ModuleImportFailure`,
which obviously doesn't match the `/\.Test([^.]*)$/` pattern expected
for normal tests. But in that case we just want to skip prepending the
classname and continue, not crash - this is not a fatal error.