EnumToI: also test the expression `{enum_value}.to_i.to_s`
I'm going to change the base class for Python enums to
[`enum.IntEnum`](https://docs.python.org/3/library/enum.html#enum.IntEnum),
as suggested in https://github.com/kaitai-io/kaitai_struct/issues/815.
An `enum.IntEnum` object behaves almost like an `int` in many cases, but
not all. For example, if you have a enum def like this:

```python
class Animal(enum.IntEnum):
    DOG = 4
    CAT = 7
```

...then `Animal.DOG == 4` is `True`. Consequently, if you do
`self.assertEqual(Animal.DOG, 4)` in a unit test, it will pass. However,
the result of `str(Animal.DOG)` depends on the version of Python:

- before Python 3.11: `str(Animal.DOG) == 'Animal.DOG'`
- since Python 3.11: `str(Animal.DOG) == '4'`

So at least as long as we support older Pythons than 3.11, a KS
expression `animal::dog.to_i.to_s` has to be translated as
`str(int(Animal.DOG))` (not just `str(Animal.DOG)`, because that won't
yield the correct result `'4'` in Python 3.10 and older). Therefore,
this commit adds an assertion to protect against this potential mistake.