cpp_builder: add `cpp_stl_11/gcc11` log as test (failed recovery)
See https://github.com/kaitai-io/ci_targets/pull/10#issue-1783059224:

> * cpp_stl_11 on gcc11 could be added too, but currently results in an
>   error that I will address separately, so commented out for now.

So this language/target combination is currently disabled in the CI, but
it's possible to run it locally as `./docker-ci -t cpp_stl -u _11 -i
gcc11`. This is also how the log added in this commit was created.

For reference, the CppBuilder's recovery failure looks like this:

```
#### CppBuilder: process_status: #<Process::Status: pid 15 exit 2>
#### CppBuilder: build attempt 1: elapsed: 405s
#### CppBuilder: build failed
#### CppBuilder: errors detected in bogus files:
#### CppBuilder: - "/usr/include/c++/11/type_traits"
#### CppBuilder: unable to recover, bailing out :(
```