cpp_builder: support multiline "In file included from ..." blocks
Note that although we're now able to read and correctly classify the
lines associated with the "In file included from ..." block (and of
course each file is on its own line), we still only keep the last file
entry. So far, this seems to be sufficient for our purposes, because the
.cpp file that we're interested in always seems to be the last file, at
least in the GCC 11 log (see
`builder/spec/cpp/gcc_h/test_out/cpp_stl_11/build-1.log`).

If it turns out that we need something better, of course we can always
improve it.