docker-ci: fix Docker image availability check on Windows
The old check `docker images -q "$IMG_NAME" | grep -q '[0-9a-f]'` is
prone to false positives - it may think the image is available when it's
not.

In particular, this happens with Docker Desktop 4.27.2 on Windows 10. If
the output of `docker images -q "$IMG_NAME"` is piped to another
command, it produces `failed to get console mode for stdout: The handle
is invalid.`. This is then processed by `grep`, which finds some
characters (`[a-f]`) it wants, so it exits with success.

To prevent this, I added the `2> /dev/null` redirect (which I'm not sure
helps, but people on the internet use it too) and also the `--no-trunc`
option to `docker images`, which changes the output format of hashes to
`sha256:[0-9a-f]+` format. This can be matched with greater confidence
using `grep`.