aggregate/compare_with_prev: show `failure` changes
Previously, all changes keeping the same `status` (especially if the new
status of a `failed` test is still `failed`) were ignored. This is
sometimes inconvenient, because we want to know when `failure.message`
or `failure.trace` changes, since that may indicate that a different
error is happening now than before (either the old error has been fixed
but now further in the code another error occurs, or a new error has
been introduced that occurs before the existing one - we want to know
about both cases).