Remove unused `imports` from KST specs
See the source code of the KST translator - the `imports` list in .kst
specs is copied to the `extraImports` property of the `TestSpec` case
class, and `extraImports` are only used to generate `import` statements
into generated test specs in C++, Java, Nim and Ruby. This means
`import` statements in test specs don't need to duplicate
`meta/imports`, it's enough to have extra imports only for specs that
are referenced from expressions in `asserts` (for example, `imports:
enum_0` is needed in `enum_import.kst` because there is an assertion
with `expected: enum_0::animal::cat`).

Since `enum_import.kst` is the only spec that refers from `asserts` to
any of the specs listed in `imports`, this commit removes the
unnecessary `imports` from all other .kst specs.

This partly addresses
<https://github.com/kaitai-io/kaitai_struct/issues/1099>. While this
doesn't solve the root of the problem, at least now tests that import
specs that are not tested directly and also not imported from any other
spec (or if they are, then only from a test that comes later in the
alphabet and thus has not run yet) no longer pass. This means that with
the current state of KSC, this breaks 4 tests in Ruby:

* EnumToIClassBorder1
* ImportsAbs
* ImportsCircularA
* ImportsRel1

The fact that these tests now fail is a good thing, because ultimately,
we want to ensure that all generated Ruby format modules contain the
`require` statements they themselves need to function, which is not the
case right now, so they *should* be failing.

Ironically, I believe that the `imports` in KST specs removed in this
commit were in place specifically to make Ruby tests pass (because
AFAIK, they are unnecessary from the perspective of any other language).
However, this reliance of Ruby format modules on external `require`s is
incorrect.