KST translator: log "... generating $outFile" even if it fails
Previously, in the case of a `NotImplementedError` thrown when
translating KST specs with the `exception` key to some languages (namely
Construct, Nim and Rust), the full stack trace of the
`scala.NotImplementedError: an implementation is missing` error was
printed to the console 3 times when using `./spec_kst_to_all -t all
...`, but it was not clear from the output for which targets it actually
occurred. To fix this, this commit ensures that the "... generating
$outFile" message is printed even if the generation fails (not just on
success) - the "... generating $outFile" message right before the error
in the console output then tells exactly which spec failed to be
generated.