Rust: don't suppress the `overflowing_literals` error
This breaks the YamlInts test, which is correct because it should be
failing. All of the checked value instances (`test_{u4,u8}_{dec,hex}`)
have type `i32` in the `compiled/rust/yaml_ints.rs` file generated by
KSC, but that's wrong because none of the integer constants that these
value instances should contain according to `formats/yaml_ints.ksy` fit
within the `i32` range (which the Rust compiler, i.e. `rustc`, correctly
reports as long as the `overflowing_literals` lint is enabled, which is
the default). So the fact that this test now fails reflects reality.