Go: migrate to modules to support Go 1.22
See https://github.com/kaitai-io/kaitai_struct/issues/1083

This commit migrates away from the deprecated ["GOPATH development
mode"](https://go.dev/wiki/GOPATH#gopath-development-mode) to modules,
which is currently the recommended way. [This
tutorial](https://go.dev/doc/tutorial/create-module) has been used as
the primary reference during the migration.

The `prepare-go` script is added here to the `kaitai_struct_tests` repo
(which is admittedly unusual, most preparation work is traditionally
done in https://github.com/kaitai-io/kaitai_struct_docker_images or
https://github.com/kaitai-io/ci_targets) because it is necessary to run
it after every cleanup & regen of the `compiled/` directory (which is
crucial for CI as well as for local development, which is why it is
added to `ci-go`), but it will also be used when building the Docker
image in https://github.com/kaitai-io/kaitai_struct_docker_images to
make `go mod download` from `spec/go/` work.

This migration also makes it unnecessary for KSC to produce Go files to
`compiled/go/src/test_formats` (this was only meaningful for the "GOPATH
mode"), we would be perfectly happy with having `*.go` files directly in
`compiled/go/`. Therefore, I think this should be changed in KSC, but
that's obviously a breaking change and not needed for the migration in
this commit, so let's leave that as a future improvement.