aggregate/compare_with_prev: fail on non-existent filenames
Until now, the script has accepted non-existent filenames as if the
files were empty.

After some time, I discovered that this is not the best idea - if you
run this script locally and make a mistake when typing the paths of one
of the files, you get a multi-screen diff between the non-existing and
existing file, which is useless (I always had two existing files in
mind, but I made a typo in one of the paths).

This behavior existed only to allow new CI targets that have no history
in the https://github.com/kaitai-io/ci_artifacts repo (or if the
previous `ci.json` report couldn't be downloaded for some other reason,
but that shouldn't normally happen) to finish successfully. However, I
believe I've found a better way to handle this, which is introduced in
this commit.