str_encodings_escaping_enc.ksy: improve comment
Using `value` as a key name in the YAML example might have been
confusing to some, because in Kaitai Struct, the `value` key is
traditionally used for expressions in value instances. However, such an
interpretation wouldn't fit here, because `here ' is a single quote` is
clearly not a well-formed KS expression. A different key name like `foo`
should prevent this potential confusion.