Revert "ExprBytesOps: also test the `bytes.to_s(encoding)` method"
This reverts commit 1d8116f2ff1f39b3729b9d62fd853109e962d026.

I'm reverting this for "compatibility reasons": it turned out that the
U+00A0 character was causing chaos in a number of target languages and
our CI infrastructure, which wasn't really the point. I've described the
cause in https://github.com/kaitai-io/kaitai_struct/issues/1133. I will
probably cover this case in a separate test sometime in the future.