JavaScriptSG: simplify byte array equality assertions
Given that all byte arrays are `Uint8Array`s since
https://github.com/kaitai-io/kaitai_struct_compiler/commit/ec064e3930e03f74df50359dc89836e778a48b2d,
we don't need to compare the byte arrays indirectly via the `hexString`
helper. The built-in `assert.deepStrictEqual()` function will do just as
well (and unlike `hexString`, it requires both arguments to be of the
same type, which is a good thing).

The only difference is that when using `assert.deepStrictEqual()`,
Node.js will output the bytes *in decimal* if the assertion fails.
Perhaps there is a way to customize that so that it would use
hexadecimal, but I don't think it matters.

For example, here's what the assertion failure message looks like:

```console
$ node
Welcome to Node.js v20.14.0.
Type ".help" for more information.
> const assert = require('assert');
undefined
> assert.deepStrictEqual(new Uint8Array([0x41, 0x42]), new Uint8Array([0x41, 0x62]));
Uncaught AssertionError [ERR_ASSERTION]: Expected values to be strictly deep-equal:
+ actual - expected

  Uint8Array(2) [
    65,
+   66
-   98
  ]
```