JavaScript: rename the `{Format}` parameter to `{Format}_`
... also in the remaining specs that haven't yet been regenerated in the
previous commit.

This is to indicate that the meaning of this parameter changed in the
previous commit because of the change to
`helpers/javascript/testHelper.js`. Previously, the test body received
the parser constructor function directly, but now it only receives the
wrapper object that contains the constructor under the name of the
format as a key. Appending an underscore `_` is our own convention to
indicate this, see
https://github.com/kaitai-io/kaitai_struct_compiler/blob/cc4e73b89bf120e57a6577dd70f72a4abd3d25e8/shared/src/main/scala/io/kaitai/struct/languages/JavaScriptCompiler.scala#L662-L666

https://github.com/kaitai-io/kaitai_struct/issues/1074 is related.