JavaScript: use `assert.strictEqual` in manually written specs
[`assert.equal`](https://nodejs.org/docs/latest-v22.x/api/assert.html#assertequalactual-expected-message)
uses the loose equality `==` operator, which is generally discouraged.
It performs various type conversions if the operand types are different,
which is usually not desirable. The strict equality `===` operator used
by `assert.strictEqual` is preferable because it always returns `false`
if the operand types are not the same.

See also
<https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness#loose_equality_using>:

> In most cases, using loose equality is discouraged. The result of a
> comparison using strict equality is easier to predict, and may
> evaluate more quickly due to the lack of type coercion.