Lua: assume that the `lua` executable is in `PATH`
It makes more sense to add `lua` to `PATH` in the Docker image and have
scripts like `ci-lua` rely on that, rather than forcing every calling
script to know where Lua is located. Therefore, `ENV
PATH="$PATH:/lua_install/bin"` was added to Lua image's `Dockerfile` in
https://github.com/kaitai-io/kaitai_struct_docker_images/commit/88807ee0d4a771a650ad333958c674e035cfa0e9,
so `lua` is now globally available in the image.