cpp_builder: sort files in project file on Ruby 2.7 and older
This should make the build logs easier to follow because it ensures that
the source files will be compiled (roughly, it may not be perfect if the
compilation is parallelized) in alphabetical order, not in chaotic order
in which Ruby writes source files into the project file
`disposable.cmake` on Ruby < 2.7 before this commit. For example, notice
the first three compiled test files in the build log of our
`cpp_stl_11/gcc4.8-linux-x86_64` CI target
(https://github.com/kaitai-io/ci_artifacts/blob/da8385e29709e69da3ae13806aac41758e9fe34f/test_out/cpp_stl_11/build-1.log#L20-L23):

```
[  0%] Building CXX object CMakeFiles/ks_tests.dir/test_repeat_until_calc_array_type.cpp.o
/usr/bin/c++ (...) -c /tests/spec/cpp_stl_11/prereq/custom_fx_no_args.cpp
[  0%] Building CXX object CMakeFiles/ks_tests.dir/test_docstrings_docref.cpp.o
[  1%] Building CXX object CMakeFiles/ks_tests.dir/test_expr_mod.cpp.o
```

This is obviously quite far from alphabetical order -
`test_repeat_until_calc_array_type` is closer to the end of the
alphabet, whereas `test_docstrings_docref` is near the beginning.

The reason Ruby 2.7 and older yield the
files in chaotic order (as we saw above) is due to the fact that the
order in which `Dir.glob` returns filenames is system-dependent, see
<https://ruby-doc.org/2.7.8/Dir.html#method-c-glob>:

> Case sensitivity depends on your system (File::FNM_CASEFOLD is
> ignored), as does the order in which the results are returned.

This changed in Ruby 3.0 - see
<https://ruby-doc.org/3.0.7/Dir.html#method-c-glob>:

> The results which matched single wildcard or character set are sorted
> in binary ascending order, unless false is given as the optional
> `sort` keyword argument.