valgrind_xml_parser: assign at most one test to each `<error>`
The existing approach that allowed attributing a single `<error>` to
multiple tests didn't make sense. If two tests trigger the same memory
leak (for example because they import the same generated parser), there
will be two `<error>` elements, one for each instance. It cannot really
happen that some error would start in one `test_*` test specification
and end up in another within the same stack trace, because each test
specification is an independent entry point (they don't call each
other).

On a related note, the generated parsers (`compiled/cpp_stl*/*.cpp`) are
no longer used to extract the test name. It turns out that every
`<stack>` has a `<frame>` with a test spec (`spec/cpp_stl*/test_*.cpp`),
so we can rely on that. If this is not the case for some reason, we
print a warning so that such an `<error>` element is not silently
ignored.

Also, we only look into the first `<stack>` element now, as opposed to
searching all of them. In practice, there are at most two `<stack>`
elements: the second `<stack>` is only present in `InvalidRead` or
`InvalidFree` errors (at least in our history of CI runs) after an
`<auxwhat>` element that says "Address ... is ... bytes after a block of
size ... alloc'd" or "... inside a block of size ... free'd". This other
`<stack>` only shows where this block was allocated or freed, which
doesn't seem to be relevant in finding the test responsible for the
error and we shouldn't have to use it.