ToStringCustom: use `str()` (not `repr()`) in Python
See https://github.com/kaitai-io/kaitai_struct/issues/732

I am convinced that in this case it is more idiomatic to use `__str__()`
than `__repr__()`. See for example
https://stackoverflow.com/a/2626364/12940655 and
https://realpython.com/python-repr-vs-str/. `__repr__` is intended for
the programmer and should be unambiguous. So for objects, it almost
always includes the name of the class - for example,
`repr(datetime.date.today())` returns `'datetime.date(2025, 8, 14)'`.
This also corresponds to what is stated at
https://docs.python.org/3.13/library/functions.html#repr:

> For many types, this function makes an attempt to return a string that
> would yield an object with the same value when passed to `eval()`;
> (...)

In contrast, `__str__` is not intended to be unambiguous. It should
provide the "informal" or nicely printable string representation of an
object, intended for the user. This much better fits the purpose of
`to-string`, as proposed in
https://github.com/kaitai-io/kaitai_struct/issues/732