aggregate/compare_with_prev: pretty-print, use `git diff`
This makes the comparison much easier to read. Previously, especially
with "changed" tests, it was almost impossible to decipher whether there
had been a real change in the test result (for example, a different
error occurred in the previous run than now), or whether it was the same
error and only some line numbers had changed. This was because of the
unfortunate, utterly unreadable format of `{... previous super long
minified "JSON" dump ...} -> {... new super long minified "JSON" dump
...}`, where the dumps can often be 1000-2000 characters long because
they encode complete stack traces.

Now, `git diff` will clearly show a minimal diff of what exactly
changed. If only a line number in the stack trace changed, we'll be
presented with a short diff with just that one-line change. This is also
enabled by the `#pretty_inspect` method, which seems to do a pretty good
job at deciding where to put newlines for decent readability, or in the
case of short dumps, to leave them on a single line.