{Java,Python} specwrite: add CastToTop.InstDisabled
This new test case demonstrates the problem pointed out in
https://github.com/kaitai-io/kaitai_struct/issues/1245#issuecomment-3148564831:
disabling the "to write" flag of an instance has no effect on the
`_fetchInstances()` / `_fetch_instances()` method. So
`_fetchInstances()` still expects disabled instances to be set, which
doesn't make sense and in fact makes it impossible for this new test
case to be successfully serialized.