zig_json_parser: fix crash on invalid UTF-8 in stderr.log
For the record, the crash occurred after implementing the `bytesToStr`
method in the Zig runtime library simply as an identity function, i.e.
always returning the input bytes. This is exactly how the
`KS_STR_ENCODING_NONE` mode in the C++ runtime library works:
https://github.com/kaitai-io/kaitai_struct_cpp_stl_runtime/blob/c7009877d4b0ca804442e796c996eb4501f3203d/kaitai/kaitaistream.cpp#L983-L986

It can be argued that this is not really a correct implementation of
`bytes_to_str`, but that's beside the point - the
`aggregate/convert_to_json` script should be robust and shouldn't crash
even in this case.